// wdp-be/src/commons/enums/order.enum.ts
export enum ORDER_TYPES {
  READY = 'READY',
  PREORDER = 'PREORDER',
  PRESCRIPTION = 'PRESCRIPTION',
}

export enum ORDER_STATUS {
  PENDING = 'PENDING',
  PENDING_PAYMENT = 'PENDING_PAYMENT',
  PAID = 'PAID',
  PROCESSING = 'PROCESSING',
  CONFIRMED = 'CONFIRMED',
  SHIPPED = 'SHIPPED',
  DELIVERED = 'DELIVERED',
  RETURNED = 'RETURNED',
  CANCELLED = 'CANCELLED',
}